Link styling: underlines/opacity
Links look much better when they become 70% opaque when hovered over, and I added two types of underlines, which was surprisingly hard. For the menu ones, I could add a small blue underline, but for the headers, it needed to be a border-bottom, which then had to be subtracted from the menu ones.